Block 4 - Shell History (#14)
This adds support for a very basic shell history.

By default, the history is stored in `.history` in the current directory.
This is, of course, not ideal for many reasons, but it's the lowest
common denominator across different operating systems.

The history file location can be overridden by setting `HISTORY_PATH`
at least. In a real-world scenario, we would store the history in the home
directory or in a config-file location, but this is very much dependent
on the operating system. Since this is foremost a Rust course,
let's keep it simple and not get too hung up on such details.